feat: reorganize utility imports and add shared module for cross-page table merging
