+++ mineru/backend/office/model_output_to_middle_json.py
@@ -1,7 +1,7 @@
 import re
 from collections import defaultdict
 
-from mineru.backend.html_image_utils import replace_inline_table_images, save_span_image_if_needed
+from mineru.backend.utils.html_image_utils import replace_inline_table_images, save_span_image_if_needed
 from mineru.backend.office.office_magic_model import MagicModel
 from mineru.utils.enum_class import BlockType
 from mineru.version import __version__
