+++ mineru/backend/utils/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) Opendatalab. All rights reserved.
+
+from .shared import cross_page_table_merge, exclude_progress_bar_idle_time
+
+__all__ = [
+    "cross_page_table_merge",
+    "exclude_progress_bar_idle_time",
+]
