+++ mineru/backend/utils/html_image_utils.py
