+++ mineru/backend/utils/shared.py
