+++ mineru/backend/vlm/model_output_to_middle_json.py
@@ -5,8 +5,8 @@ import numpy as np
 from loguru import logger
 from tqdm import tqdm
 
-from mineru.backend.html_image_utils import replace_inline_table_images
-from mineru.backend.para_block_utils import (
+from mineru.backend.utils.html_image_utils import replace_inline_table_images
+from mineru.backend.utils.para_block_utils import (
     build_para_blocks_from_preproc,
     cleanup_internal_para_block_metadata,
     merge_para_text_blocks,
@@ -27,7 +27,7 @@ if llm_aided_config:
     title_aided_config = llm_aided_config.get('title_aided', {})
     if title_aided_config.get('enable', False):
         from mineru.utils.llm_aided import llm_aided_title
-        from mineru.backend.ocr_det_utils import (
+        from mineru.backend.utils.ocr_det_utils import (
             detect_ocr_boxes_from_padded_crop,
             get_ch_lite_ocr_det_model,
         )
